pyproject.toml
[project]
name = "check-ia"
version = "0.1.0"
description = "AI-powered fact-checking platform for the French-speaking Sahel"
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4==4.12.3",
    "celery==5.4.0",
    "deep-translator==1.11.4",
    "django==5.1.1",
    "django-cors-headers==4.4.0",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.3.1",
    "fr-core-news-sm",
    "huggingface-hub==0.25.1",
    "numpy>=1.24.3,<2.0",
    "openai>=1.102.0",
    "pillow==10.0.0",
    "psycopg2-binary==2.9.9",
    "python-dotenv==1.0.1",
    "redis==5.1.1",
    "requests==2.32.3",
    "spacy==3.8.2",
    "supabase==2.11.0",
    "torch==2.2.2",
    "tqdm==4.66.5",
    "transformers==4.45.1",
    "gunicorn==22.0.0",
    "whitenoise==6.7.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings_test"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.uv.sources]
fr-core-news-sm = { url = "https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.8.0/fr_core_news_sm-3.8.0-py3-none-any.whl" }
